테스트 코드 리팩토링
작성된 테스트 코드를 외부로 빼는 작업을 진행함.
