+++ src/main/java/com/example/testdrivenstudy/product/DiscountPolicy.java
@@ -1,0 +1,5 @@
+package com.example.testdrivenstudy.product;
+
+enum DiscountPolicy {
+    NONE
+}
