+++ src/main/java/com/example/testdrivenstudy/product/ProductAdapter.java
@@ -1,0 +1,14 @@
+package com.example.testdrivenstudy.product;
+
+class ProductAdapter implements ProductPort {
+    private final ProductRepository productRepository;
+
+    ProductAdapter(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Override
+    public void save(final Product product) {
+        productRepository.save(product);
+    }
+}
