+++ src/main/java/com/example/testdrivenstudy/product/ProductPort.java
@@ -1,0 +1,5 @@
+package com.example.testdrivenstudy.product;
+
+interface ProductPort {
+    void save(final Product product);
+}
