+++ src/main/java/com/example/testdrivenstudy/product/ProductService.java
@@ -1,0 +1,15 @@
+package com.example.testdrivenstudy.product;
+
+class ProductService {
+    private final ProductPort productPort;
+
+    public ProductService(ProductPort productPort) {
+        this.productPort = productPort;
+    }
+
+    public void addProduct(final AddProductRequest request) {
+        final Product product = new Product(request.name(), request.price(), request.discountPolicy());
+
+        productPort.save(product);
+    }
+}
