+++ src/test/java/com/example/testdrivenstudy/product/ProductServiceTest.java
@@ -1,16 +1,14 @@
 package com.example.testdrivenstudy.product;
 
-import org.apache.catalina.Store;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.util.Assert;
 
-import java.util.HashMap;
-import java.util.Map;
-
-/*변수에 final 키워드 변경이 필요 없는 변수를 불변으로 만듬
+/*
+변수에 final 키워드 변경이 필요 없는 변수를 불변으로 만듬
 변수의 값이 변경되는 오류를 방지하는 차원에서 값이 변경되지 않는 모든 변수를 불변으로 만듭니다.
-무분별하게 setter를 만들지 않는것과 비슷.!*/
+무분별하게 setter를 만들지 않는것과 비슷.!
+단축키 select refactoring F6
+*/
 
 class ProductServiceTest {
     private ProductService productService;
@@ -24,84 +22,26 @@ class ProductServiceTest {
         productService = new ProductService(productPort);
     }
 
-    private class ProductService {
-        private final ProductPort productPort;
-
-        public ProductService(ProductPort productPort) {
-            this.productPort = productPort;
-        }
-
-        public void addProduct(final AddProductRequest request) {
-            final Product product = new Product(request.name(), request.price(), request.discountPolicy());
-
-            productPort.save(product);
-        }
-    }
-
     @Test
     void 상품등록(){
-        final String name = "상품명"; //command + alt + v 자동완성
+        //command + alt + v 자동완성
+        /*
+        final String name = "상품명";
         final int price = 1000;
-
-        DiscountPolicy discountPolicy = DiscountPolicy.NONE;
+        final DiscountPolicy discountPolicy = DiscountPolicy.NONE;
         final AddProductRequest request = new AddProductRequest(name, price, discountPolicy);
+        final 변수에서 command alt M 으로 밖으로 뺌 -> 상풍등록요청_생성()
+        */
+        final AddProductRequest request = 상풍등록요청_생성();
         productService.addProduct(request);
     }
-    /**
-     * @param /alt+Enter 자동 바인드
-     */
-    private record AddProductRequest(String name, int price, DiscountPolicy discountPolicy) {
-        private AddProductRequest {
-            Assert.hasText(name, "상품명은 필수입니다.");
-            Assert.isTrue(price > 0, "상품 가격은 0보다 커야 합니다.");
-            Assert.notNull(discountPolicy, "할인 정책은 필수입니다.");
-        }
 
-        }
-    private enum DiscountPolicy {
-        NONE
-    }
-    private class Product {
-        private Long id;
-        private final String name;
-        private final int price;
-        private final DiscountPolicy discountPolicy;
-        public Product(String name, int price, DiscountPolicy discountPolicy) {
-            Assert.hasText(name, "상품명은 필수입니다.");
-            Assert.isTrue(price > 0, "상품 가격은 0보다 커야 합니다.");
-            Assert.notNull(discountPolicy, "할인 정책은 필수입니다.");
-            this.name = name;
-            this.price = price;
-            this.discountPolicy = discountPolicy;
-        }
-        public void assignId(Long id) {
-            this.id = id;
-        }
-        public Long getId() {
-            return id;
-        }
+    private static AddProductRequest 상풍등록요청_생성() {
+        final String name = "상품명";
+        final int price = 1000;
+        final DiscountPolicy discountPolicy = DiscountPolicy.NONE;
+        final AddProductRequest request = new AddProductRequest(name, price, discountPolicy);
+        return request;
     }
 
-    private interface ProductPort {
-        void save(final Product product);
-    }
-
-    private class ProductAdapter implements ProductPort {
-        private final ProductRepository productRepository;
-        private ProductAdapter(ProductRepository productRepository) {
-            this.productRepository = productRepository;
-        }
-        @Override
-        public void save(final Product product) {
-            productRepository.save(product);
-        }
-    }
-    private class ProductRepository {
-        private Map<Long, Product> persistence = new HashMap<>();
-        private Long sequence = 0L;
-        public void save(Product product) {
-            product.assignId(++sequence);
-            persistence.put(product.getId(), product);
-        }
-    }
 }
